requirements.txt
python-dotenv

# video_summary_bot_core
yt-dlp==2024.3.10
openai-whisper==20231117
pydub==0.25.1

# audio_summary_simple_bot
openai

# telegram bot
aiogram==3.4.1
